requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Database
sqlite3  # Built into Python

# Resume parsing
PyMuPDF==1.23.8  # For PDF parsing
python-docx>=0.8.11  # For Word documents (optional)

# RAG and ML
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.24.3
torch>=2.0.0  # Required by sentence-transformers

# Additional ML dependencies
transformers>=4.35.0  # For advanced NLP
tf-keras>=2.15.0  # For TensorFlow compatibility

# HTTP requests and utilities
requests==2.31.0
python-dateutil==2.8.2

# Optional: For enhanced text processing
nltk>=3.8.1
spacy>=3.7.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1

# New dependencies for multi-modal processing
opencv-python==4.8.1.78
moviepy==1.0.3
SpeechRecognition==3.10.0
pyaudio==0.2.11

# Additional dependencies
pandas==1.5.3
Pillow==10.1.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
jinja2==3.1.2
aiofiles==23.2.1
httpx==0.25.2
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1
pre-commit==3.5.0
coverage==7.3.2
pytest-cov==4.1.0
bandit==1.7.5

# Optional: AI Services for real LLM integration
# Free alternatives (recommended):
transformers>=4.35.0  # Local AI models (FREE)
torch>=2.0.0  # Required for transformers
groq  # Groq API client (FREE tier available) - install separately if needed

# Paid alternatives (optional):
# openai>=1.0.0  # For OpenAI API integration (PAID)
# anthropic>=0.7.0  # For Anthropic Claude API (PAID)

# Optional: Enhanced email support
secure-smtplib>=0.1.1  # Enhanced SMTP security

# Optional: Environment management
python-decouple>=3.8  # Better environment variable handling 